device-management/requirements.txt
FastAPI==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pika==1.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyJWT==2.8.1
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
httpx==0.25.2
